cant delete if not 0 stock

+++ app/Controllers/product-delete.php
@@ -15,18 +15,21 @@ if($_SERVER['REQUEST_METHOD'] == "POST" && $row)
 	$_POST['source'] = "Products";
 	$_POST['action'] = "DELETE";
 
-	//$product->audit_trail($row['id'], null, "Products", "DELETE");
-	$product->audit_trail($row['id'], $_POST);
-	$product->delete_item($row['id']);
-  	
+	$db = new Database();
+	$result = $db->query("SELECT stock FROM products WHERE id = $id");
+	if ($result[0]['stock'] == 0){
+		$product->audit_trail($row['id'], $_POST);
+		$product->delete_item($row['id']);
+		redirect('admin&tab=products');
+	} else {
+		echo '<script>alert("You can only delete a product if the stock is 0");</script>';
+	}
 	/*/delete old image
 	if(file_exists($row['image']))
 	{
 		unlink($row['image']);
 	}
 	*/
-	redirect('admin&tab=products');
-	
 }
